ecommerce/requirements.txt
asgiref==3.4.1
astroid==2.8.4
autopep8==1.4.4
certifi==2019.6.16
cffi==1.15.0
chardet==3.0.4
colorama==0.4.4
cryptography==35.0.0
defusedxml==0.6.0
Django==3.2.8
django-allauth==0.44.0
django-countries==5.3.3
django-crispy-forms==1.13.0
django-debug-toolbar==1.10.1
django-extensions==3.1.3
idna==2.8
isort==5.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.1
pycodestyle==2.5.0
pycparser==2.20
PyJWT==2.3.0
pylint==2.11.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.16.0
sqlparse==0.2.4
stripe==2.27.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.24.3
wrapt==1.11.2
